change atomic int to atomic size_t in gameRemaining

+++ ex3/ex3/BattleshipTournamentManager.h
@@ -16,7 +16,7 @@ struct Round
 {
 
 	int roundNumber;
-	std::atomic<int> numOfGamesLeft;
+	std::atomic<size_t> numOfGamesLeft;
 	bool status;
 	//std::mutex
